+++ doc/CHANGES.rst
@@ -13,3 +13,5 @@ CHANGELOG
 
 0.1.0
 -----
+
+* [ADD] **Jsonb** column